feat: send back requests
This commit is a huge one. It changes three key things:
1. Responses now contain the stream to write to (ownership taken from
the Request's BodyReader)
2. The stream contained within Requests is now in a Box rather than
Arc<Mutex>> because the request will be processed by middlewares and
endpoints sequentially, not in parallel
3. Handlers now own the Request and signal to the caller whether to
continue to the next handler with an enum. This will be used to
implement middleware

Introduces new APIs:
- SyncableStream: a trait for streams to abstract away the concrete
stream from Request and Response
- ResponseBuilder: a builder to make creating Responses easier