fix: correct JSON parsing in CRUD example
This is verbose and annoying. I need a helper method for this. The
problem is that the way I've implemented BodyReader means I can't
propagate the desired type down to its read_json because BodyReader is
used as a trait object, meaning it can't have generics. I'll need to
rethink this structure

Anyway, the bug was using to_string instead of as_str. See
https://docs.rs/serde_json/latest/serde_json/#operating-on-untyped-json-values