feat: add graceful shudown + CTRL+C shutdown
Adds the ability to close the server with the shutdown method on
HTTPListener or pressing CTRL + C in the terminal

Implementation details:
- Uses an AtomicBool to avoid locking on every request to check the
shutdown signal
- My RequestQueue doesn't have close semantics because it's a queue
rather than a channel. I just push a shutdown message for each thread
into the queue telling them to join
- To get the TCPStream::accept loop to stop blocking, I send a dummy
request to get it to re-evaluate the shutdown signal. This is ugly but
seems like the only solution apart from non-blocking mode. I chose to
not go with that solution as it would add complexity due to dealing with
platform-specific APIs for polling for a file descriptor. I will add
non-blocking mode in the future